feat: 프로젝트 종합 대시보드 UI 컴포넌트 구현 및 파일 구조 변경
    - 프로젝트 목록 테이블 공통 컴포넌트 및 페이지네이션 컴포넌트 구현
    - 검색 기능을 포함한 서치바(인풋 + 검색 버튼) 공통 컴포넌트 추가
    - 프로젝트 현황을 보여주는 StatusCard 및 StatusCards 컴포넌트 구현
    - types, data 폴더 위치 변경
    - app안의 components폴더를 루트components폴더의common폴더로 이전후 삭제
