--- app/projects/page.tsx
@@ -1,3 +1,0 @@
-import ProjectsPage from "@/components/pages/ProjectsPage";
-
-export default ProjectsPage;
