+++ components/common/SearchSection.tsx
