+++ components/common/StatusCard.tsx
