+++ types/index.ts
@@ -1,0 +1,2 @@
+export * from './projects';
+export * from './pagination';
