+++ types/pagination.ts
@@ -1,0 +1,6 @@
+export interface PaginationProps {
+  currentPage: number;
+  totalCount: number;
+  pageSize: number;
+  onPageChange: (page: number) => void;
+}
